+++ Models/AppDataGooglePlaySearchOrganicSerpElementItem.cs
@@ -30,7 +30,7 @@ namespace DataForSeo.Client.Models
         /// the total number of reviews of the app
         /// </summary>
         [JsonProperty("reviews_count", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public double? ReviewsCount { get; set; }
+        public long? ReviewsCount { get; set; }
 
         /// <summary>
         /// indicates whether the app is free
