+++ Models/BacklinksBulkReferringDomainsLiveItem.cs
@@ -19,26 +19,26 @@ namespace DataForSeo.Client.Models
         /// <br/>note that we calculate main domains (root domains, like example.com) and their subdomains (e.g. blog.example.com) separately for this metric
         /// </summary>
         [JsonProperty("referring_domains", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public double? ReferringDomains { get; set; }
+        public long? ReferringDomains { get; set; }
 
         /// <summary>
         /// number of domains pointing at least one nofollow link to the target
         /// </summary>
         [JsonProperty("referring_domains_nofollow", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public double? ReferringDomainsNofollow { get; set; }
+        public long? ReferringDomainsNofollow { get; set; }
 
         /// <summary>
         /// number of referring main domains pointing to the target
         /// <br/>the number of primary (root) domains referring to your target
         /// </summary>
         [JsonProperty("referring_main_domains", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public double? ReferringMainDomains { get; set; }
+        public long? ReferringMainDomains { get; set; }
 
         /// <summary>
         /// number of main domains pointing at least one nofollow link to the target
         /// </summary>
         [JsonProperty("referring_main_domains_nofollow", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public double? ReferringMainDomainsNofollow { get; set; }
+        public long? ReferringMainDomainsNofollow { get; set; }
 
         private IDictionary<string, object> _additionalProperties;
