+++ Models/BaseBusinessDataSerpElementItem.cs
@@ -28,13 +28,13 @@ namespace DataForSeo.Client.Models
         /// <br/>positions of elements with different type values are omitted from the rank_group
         /// </summary>
         [JsonProperty("rank_group", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public double? RankGroup { get; set; }
+        public int? RankGroup { get; set; }
 
         /// <summary>
         /// absolute rank among all the elements
         /// </summary>
         [JsonProperty("rank_absolute", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public double? RankAbsolute { get; set; }
+        public int? RankAbsolute { get; set; }
 
         private IDictionary<string, object> _additionalProperties;
